Update Nuget Package, Remove MachineManagerService.Amqp project
