+++ CargoEdu.sln
@@ -21,8 +21,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "R7P.MachineManagerService.D
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "R7P.MachineManagerService.Infrastructure.EntityFramework", "src\MachineManagerService\Infrastructure\Infrastructure.EntityFramework\R7P.MachineManagerService.Infrastructure.EntityFramework.csproj", "{8735007C-F29C-4540-9B82-50C773B2CED7}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "R7P.MachineManagerService.AMQP", "src\MachineManagerService\Amqp\R7P.MachineManagerService.AMQP.csproj", "{BA7244B5-9DF9-457F-AF92-67C14E49092D}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Domain", "Domain", "{E46D65B5-18C9-44E7-9E29-FC6268463509}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Infrastructure", "Infrastructure", "{A6DC297A-0CDF-4077-95B2-5C4CEA25646D}"
@@ -65,7 +63,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "R7P.MachineManagerService.W
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "MainGateService", "MainGateService", "{41810905-04CC-4841-8220-9FF832D42F6B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "R7P.SharedModels", "src\SharedModels\R7P.SharedModels\R7P.SharedModels.csproj", "{55B6937B-F205-4A4F-8163-355C2045A205}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "R7P.SharedModels", "src\SharedModels\R7P.SharedModels\R7P.SharedModels.csproj", "{55B6937B-F205-4A4F-8163-355C2045A205}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SharedModels", "SharedModels", "{BAEB9064-F4C7-47D8-9E9C-EE4E7595C15C}"
 EndProject
@@ -103,10 +101,6 @@ Global
 		{8735007C-F29C-4540-9B82-50C773B2CED7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8735007C-F29C-4540-9B82-50C773B2CED7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8735007C-F29C-4540-9B82-50C773B2CED7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BA7244B5-9DF9-457F-AF92-67C14E49092D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BA7244B5-9DF9-457F-AF92-67C14E49092D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BA7244B5-9DF9-457F-AF92-67C14E49092D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BA7244B5-9DF9-457F-AF92-67C14E49092D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A3A7B0F8-6AB3-4B7F-853F-D1AD02FC5503}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A3A7B0F8-6AB3-4B7F-853F-D1AD02FC5503}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A3A7B0F8-6AB3-4B7F-853F-D1AD02FC5503}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -164,7 +158,6 @@ Global
 		{15A09201-CB38-4AF6-8312-B1FA933F230C} = {717E9A22-5E67-463B-99D4-1ADC9B84676A}
 		{2BE01C0B-0A97-4FDE-866C-AC6DE5447B21} = {E46D65B5-18C9-44E7-9E29-FC6268463509}
 		{8735007C-F29C-4540-9B82-50C773B2CED7} = {A6DC297A-0CDF-4077-95B2-5C4CEA25646D}
-		{BA7244B5-9DF9-457F-AF92-67C14E49092D} = {4246D2D8-E08A-45F3-9C73-BCFADAADF2C7}
 		{E46D65B5-18C9-44E7-9E29-FC6268463509} = {4246D2D8-E08A-45F3-9C73-BCFADAADF2C7}
 		{A6DC297A-0CDF-4077-95B2-5C4CEA25646D} = {4246D2D8-E08A-45F3-9C73-BCFADAADF2C7}
 		{4246D2D8-E08A-45F3-9C73-BCFADAADF2C7} = {195BAC64-7EFB-4F74-94D8-3DCA26910B35}
