+++ src/DeliveryCalculationService/R7P.DeliveryCalculationService.Application/R7P.DeliveryCalculationService.Application.csproj
@@ -8,7 +8,7 @@
 
 	<ItemGroup>
 		<PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="8.0.0" />
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.0" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.1" />
 		<PackageReference Include="Riok.Mapperly" Version="3.5.1" />
 	</ItemGroup>
