--- src/MachineManagerService/Amqp/Properties/launchSettings.json
@@ -1,15 +1,0 @@
-{
-  "profiles": {
-    "OrderManager": {
-      "commandName": "Project",
-      "environmentVariables": {
-        "DOTNET_ENVIRONMENT": "Development"
-      },
-      "dotnetRunMessages": true
-    },
-    "Docker": {
-      "commandName": "Docker"
-    }
-  },
-  "$schema": "http://json.schemastore.org/launchsettings.json"
-}
