+++ src/MachineManagerService/Application/Application.Application/DependencyInjection.cs
@@ -8,8 +8,10 @@ namespace R7P.MachineManagerService.Application;
 public static class DependencyInjection
 {
     public static IServiceCollection AddApplicationServices(this IServiceCollection services, IConfiguration configuration)
-    {        
+    {
+        services.AddHostedService<SchedulerService>();
         services.AddScoped<IMachineService, MachineService>();
+
 
         return services;
     }
