+++ src/MachineManagerService/Application/R7P.MachineManagerService.Application.Interfaces/R7P.MachineManagerService.Application.Interfaces.csproj
@@ -11,7 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Riok.Mapperly" Version="3.4.0" />
+    <PackageReference Include="Riok.Mapperly" Version="3.5.1" />
   </ItemGroup>
 
 </Project>
