+++ src/MachineManagerService/WebSimulator/Program.cs
@@ -12,10 +12,10 @@ namespace R7P.MachineManagerService.WebSimulator
             builder.Services.AddRazorComponents()
                 .AddInteractiveServerComponents();
 
-            //builder.Services.AddServerSideBlazor();
-
             builder.Services.AddSingleton<NotifierService>();
             builder.Services.AddSingleton<TimerService>();
+
+            builder.WebHost.UseStaticWebAssets();
 
             //string connectionString = builder.Configuration.GetSection("ConnectionStrings")["DefaultConnection"] 
             //    ?? throw new InvalidOperationException("Not found connection string");
